Mapper | Exception | Refatorando funções
Adicionado um mapper para o PetDto;
Adicionado um Exception para pet não encontrado;
Refatorando funções para utilizar as novas classes.
