+++ src/main/java/br/com/academiadev/thunderpets/exception/PetNaoEncontradoException.java
@@ -1,0 +1,8 @@
+package br.com.academiadev.thunderpets.exception;
+
+public class PetNaoEncontradoException extends Exception {
+
+    public PetNaoEncontradoException(String msg) {
+        super(msg);
+    }
+}
